requirements_preprocess.txt
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
pytesseract==0.3.13
scikit-image==0.23.2
scipy==1.13.0
flask==3.0.3
